fix login issue pear-555
